requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
fake-useragent==0.1.11
Flask==2.0.2
frozenlist==1.2.0
future==0.18.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
line-bot-sdk==2.0.1
lxml==4.7.1
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.21.5
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
